requirements.txt
Flask
SQLAlchemy
alembic
openai
python-dotenv